+++ packages/hardhat/hardhat.config.js
@@ -44,13 +44,17 @@ module.exports = {
   networks: {
     localhost: {
       url: "http://localhost:8545",
-      accounts: {
-        mnemonic: mnemonic(),
-      },
       /*
         notice no mnemonic here? it will just use account 0 of the hardhat node to deploy
         (you can put in a mnemonic here to set the deployer locally)
       */
+    },
+    l2: {
+      url: "http://localhost:8545",
+      accounts: {
+        mnemonic: mnemonic(),
+      },
+      chainId: 420
     },
     rinkeby: {
       url: "https://rinkeby.infura.io/v3/460f40a260564ac4a4f4b3fffb032dad", //<---- YOUR INFURA ID! (or it won't work)
