+++ packages/hardhat/package.json
@@ -11,7 +11,9 @@
     "eslint-plugin-prettier": "^3.1.4"
   },
   "dependencies": {
+    "@eth-optimism/contracts": "^0.1.9",
     "@eth-optimism/plugins": "^0.0.16",
+    "@eth-optimism/watcher": "^0.0.1-alpha.9",
     "@nomiclabs/hardhat-ethers": "^2.0.0",
     "@nomiclabs/hardhat-waffle": "^2.0.0",
     "@openzeppelin/contracts": "^3.2.0",
