+++ packages/react-app/src/components/Contract/FunctionForm.jsx
@@ -5,6 +5,7 @@ import React, { useState } from "react";
 import { BigNumber } from "@ethersproject/bignumber";
 import { Row, Col, Input, Divider, Tooltip, Button } from "antd";
 import { Transactor } from "../../helpers";
+import { hexlify } from "@ethersproject/bytes";
 import tryToDisplay from "./utils";
 import Blockies from "react-blockies";
 const { utils } = require("ethers");
@@ -202,6 +203,7 @@ export default function FunctionForm({ contractFunction, functionInfo, provider,
                 if (txValue) {
                   overrides.value = txValue; // ethers.utils.parseEther()
                 }
+                overrides.gasLimit = hexlify(8000000);
 
                 // console.log("Running with extras",extras)
                 const returned = await tx(contractFunction(...args, overrides));
